AI-Powered Document Intelligence Platform/backend/requirements.txt
Flask
python-dotenv
PyPDF2
langchain
pinecone-client
tiktoken
langchain-google-genai
google-generativeai
langchain-pinecone
